+++ packages/react-resizable-panels-website/src/routes/examples/Nested.tsx
@@ -20,23 +20,23 @@ function Content() {
   return (
     <div className={styles.PanelGroupWrapper}>
       <PanelGroup className={styles.PanelGroup} direction="horizontal">
-        <Panel className={styles.PanelRow} defaultSize={20}>
+        <Panel className={styles.PanelRow} defaultSize={20} minSize={10}>
           <div className={styles.Centered}>left</div>
         </Panel>
         <ResizeHandle className={styles.ResizeHandle} />
         <Panel className={styles.PanelRow} minSize={35}>
           <PanelGroup className={styles.PanelGroup} direction="vertical">
-            <Panel className={styles.PanelColumn} defaultSize={35}>
+            <Panel className={styles.PanelColumn} defaultSize={35} minSize={10}>
               <div className={styles.Centered}>top</div>
             </Panel>
             <ResizeHandle className={styles.ResizeHandle} />
-            <Panel className={styles.PanelColumn}>
+            <Panel className={styles.PanelColumn} minSize={10}>
               <PanelGroup className={styles.PanelGroup} direction="horizontal">
-                <Panel className={styles.PanelRow}>
+                <Panel className={styles.PanelRow} minSize={10}>
                   <div className={styles.Centered}>left</div>
                 </Panel>
                 <ResizeHandle className={styles.ResizeHandle} />
-                <Panel className={styles.PanelRow}>
+                <Panel className={styles.PanelRow} minSize={10}>
                   <div className={styles.Centered}>right</div>
                 </Panel>
               </PanelGroup>
@@ -44,7 +44,7 @@ function Content() {
           </PanelGroup>
         </Panel>
         <ResizeHandle className={styles.ResizeHandle} />
-        <Panel className={styles.PanelRow} defaultSize={20}>
+        <Panel className={styles.PanelRow} defaultSize={20} minSize={10}>
           <div className={styles.Centered}>right</div>
         </Panel>
       </PanelGroup>