+++ packages/react-resizable-panels-website/tests/Collapsing.spec.ts
@@ -18,7 +18,9 @@ test.describe("collapsible prop", () => {
           minSize: 10,
         }),
         createElement(PanelResizeHandle, { style: { height: 10, width: 10 } }),
-        createElement(Panel),
+        createElement(Panel, {
+          minSize: 10,
+        }),
         createElement(PanelResizeHandle, { style: { height: 10, width: 10 } }),
         createElement(Panel, {
           collapsible: true,
@@ -81,7 +83,7 @@ test.describe("collapsible prop", () => {
           minSize: 10,
         }),
         createElement(PanelResizeHandle, { style: { height: 10, width: 10 } }),
-        createElement(Panel)
+        createElement(Panel, { minSize: 10 })
       )
     );