+++ packages/react-resizable-panels-website/tests/CursorStyle.spec.ts
@@ -41,9 +41,10 @@ test.describe("cursor style", () => {
         createElement(Panel, {
           defaultSize: 50,
           id: "first-panel",
+          minSize: 10,
         }),
         createElement(PanelResizeHandle),
-        createElement(Panel, { defaultSize: 50, id: "last-panel" })
+        createElement(Panel, { defaultSize: 50, id: "last-panel", minSize: 10 })
       )
     );
   }